Added the getLastInsertedId, update, markNewsArticleAsActive, markNewsArticleAsInactive, getUserFriendlyNewsStorageName to the NewsStorageSystem interface.
The getLastInsertedId method returns the latest inserted ID in the news storage, the update, like the name implies, updates an already existent record in the database, the markNewsArticleAsActive method marks a news article as being already sent to a blog (hence the active name), the markNewsArticleAsInactive does the reverse of the previous method, and the getUserFriendlyNewsStorageName returns a user friendly name for the users to see when interacting with the application.

Also implemented the MySQLNewsStorageSystem class.